"team - rozpustenie" pridany check ci existuje team

+++ app/sp-api/app/Http/Controllers/TeamController.php
@@ -111,10 +111,13 @@ class TeamController extends Controller
         $user_id = auth()->id();
         $team = Team::whereId($team_id)->first();
 
+        if (!$team) {
+            throw new \Exception("Team doesnt exist");
+        }
+
         if ($team->user_id == $user_id){
             $team->delete();
             return response()->json('Team deleted', 200);
-
         }
 
         return response()->json('You are not owner of the team:', 301);
